+++ src/quality.rs
@@ -2,6 +2,7 @@ use strum_macros::{Display, EnumString, EnumIter, EnumCount};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
 
+/// Quality.
 #[derive(Serialize_repr, Deserialize_repr, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, Display, EnumString, EnumIter, EnumCount, TryFromPrimitive, IntoPrimitive, Clone, Copy)]
 #[repr(u32)]
 pub enum Quality {
@@ -29,44 +30,61 @@ pub enum Quality {
 }
 
 impl Quality {
+    /// Gets the related color of this quality as a hexadecimal color.
     pub fn color(&self) -> u32 {
         match self {
-            Quality::Normal => 0xB2B2B2,
-            Quality::Genuine => 0x4D7455,
-            Quality::Rarity2 => 0xFFFFFF,
-            Quality::Vintage => 0x476291,
-            Quality::Rarity3 => 0xFFFFFF,
-            Quality::Unusual => 0x8650AC,
-            Quality::Unique => 0xFFD700,
-            Quality::Community => 0x70B04A,
-            Quality::Valve => 0x56083F,
-            Quality::SelfMade => 0x70B04A,
-            Quality::Customized => 0xFFFFFF,
-            Quality::Strange => 0xCF6A32,
-            Quality::Completed => 0xFFFFFF,
-            Quality::Haunted => 0x38F3AB,
-            Quality::Collectors => 0xAA0000,
-            Quality::DecoratedWeapon => 0xFAFAFA,
+            Self::Normal => 0xB2B2B2,
+            Self::Genuine => 0x4D7455,
+            Self::Rarity2 => 0xFFFFFF,
+            Self::Vintage => 0x476291,
+            Self::Rarity3 => 0xFFFFFF,
+            Self::Unusual => 0x8650AC,
+            Self::Unique => 0xFFD700,
+            Self::Community => 0x70B04A,
+            Self::Valve => 0x56083F,
+            Self::SelfMade => 0x70B04A,
+            Self::Customized => 0xFFFFFF,
+            Self::Strange => 0xCF6A32,
+            Self::Completed => 0xFFFFFF,
+            Self::Haunted => 0x38F3AB,
+            Self::Collectors => 0xAA0000,
+            Self::DecoratedWeapon => 0xFAFAFA,
         }
     }
     
+    /// Converts a hexadecimal color into a [`Quality`].
+    /// 
+    /// # Examples
+    /// ```
+    /// use tf2_enum::Quality;
+    /// 
+    /// assert_eq!(Quality::from_color(0x8650AC).unwrap(), Quality::Unusual);
+    /// ```
     pub fn from_color(color: u32) -> Option<Self> {
         match color {
-            0xB2B2B2 => Some(Quality::Normal),
-            0x4D7455 => Some(Quality::Genuine),
-            0x476291 => Some(Quality::Vintage),
-            0x8650AC => Some(Quality::Unusual),
-            0xFFD700 => Some(Quality::Unique),
-            0x56083F => Some(Quality::Valve),
-            0x70B04A => Some(Quality::SelfMade),
-            0xCF6A32 => Some(Quality::Strange),
-            0x38F3AB => Some(Quality::Haunted),
-            0xAA0000 => Some(Quality::Collectors),
-            0xFAFAFA => Some(Quality::DecoratedWeapon),
+            0xB2B2B2 => Some(Self::Normal),
+            0x4D7455 => Some(Self::Genuine),
+            0x476291 => Some(Self::Vintage),
+            0x8650AC => Some(Self::Unusual),
+            0xFFD700 => Some(Self::Unique),
+            0x56083F => Some(Self::Valve),
+            0x70B04A => Some(Self::SelfMade),
+            0xCF6A32 => Some(Self::Strange),
+            0x38F3AB => Some(Self::Haunted),
+            0xAA0000 => Some(Self::Collectors),
+            0xFAFAFA => Some(Self::DecoratedWeapon),
             _ => None,
         }
     }
     
+    /// Converts a hexadecimal color string into a [`Quality`].
+    /// 
+    /// # Examples
+    /// ```
+    /// use tf2_enum::Quality;
+    /// 
+    /// assert_eq!(Quality::from_color_str("#8650AC").unwrap(), Quality::Unusual);
+    /// ```
     pub fn from_color_str(color: &str) -> Option<Self> {
         let len = color.len();
         let mut color = color;